refactor(handler): move schedule request-to-row mappers to schedules_request.go (#1022)
Pure code motion: --color-moved clean, no exported-API or behaviour change.
No positional comments needed fixing (moved blocks reference their siblings by
name, all of which travel to the same new file).
